Add image and some adjustment

+++ Makdempu Movie/Makdempu Movie/Entity/Review.swift
@@ -39,7 +39,7 @@ struct Review: Codable {
 struct AuthorDetails: Codable {
     let name, username: String?
     let avatarPath: String?
-    let rating: Int?
+    let rating: Float?
 
     enum CodingKeys: String, CodingKey {
         case name, username
